Move data-encoder to it's own file
